Trait pallet_schemas_rpc::SchemasApiServer

pub trait SchemasApiServer<BlockHash>: Sized + Send + Sync + 'static {
    fn get_by_schema_id(
        &self,
        schema_id: SchemaId
    ) -> RpcResult<Option<SchemaResponse>>;
    fn check_schema_validity(
        &self,
        model: Vec<u8>,
        at: Option<BlockHash>
    ) -> RpcResult<bool>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
    { ... }
}

Server trait implementation for the SchemasApi RPC API.

Required Methods

fn get_by_schema_id(
&self,
schema_id: SchemaId
) -> RpcResult<Option<SchemaResponse>>

retrieving schema by schema id

fn check_schema_validity(
&self,
model: Vec<u8>,
at: Option<BlockHash>
) -> RpcResult<bool>

validates a schema model and returns true if the model is correct.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where
BlockHash: Send + Sync + 'static + DeserializeOwned,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.